Disallow duplicated species in reactions
Since n= is supposed to be used to specify stochiometry of reactants
(or products) disallow specyfing the same reactant or product multiple
times. An alternative would be to count reactants and products and update
n= based on the result. But having both options seems redundant.

Reactions are now described in a nicer way.