Check products and reactants separately when calculating leap times
In a reaction like A→A+B, where B is very abundant, the condition for
B might be satisfied with a huge leap. But if A is scarce, we would
end up with trying to execute a reaction with an extent bigger than
the count of A. We shouldn't do that.