ResultWriterHDF5: use chunking to write out events
It turns out, that when the data that is written out
is not a multiple of the chunk size, hdf5 ignores (?) the
selection, and index errors occur. To work around the problem,
events are gathered and written in multiples of the chunk size.
Currently, the last incomplete chunk is simply lost.
It is not certain, if the data that is written is correct.