NextEventQueue: be smarter when checking if propensity was updated properly
Before, the old population and the new population were compared, and if
the it was different but the propensity didn't change, an exception was thrown.
This check is crucial, because if the propensity is recalculated without chaning,
we can assume that the propensity update algorithm isn't working properly.
But with multiple substrates, if one goes up, and the other down, propensity
might remain constant. Check this explicitly, by comparing arrays of old
and new populations in the slow path.