Do not require setting of fixedStepDt
Unless fixedStepDt was set properly, ExactStochastic would
needlessly stop every fixedStepDt. It needs only to be small
enough to enable proper reporting, can can be the whole runtime
if there is no reporting.

SteppepedStochastic does not really work if fixedStepDt is not
an even divisor of the reporting time, and will return bogus
results for the last value. Not easy to fix because log(dt)
is stored in various places.