python: round times to six decimal places
The issue was that when adding time steps the resulting number is slightly
different than the target because of rounding error. This can cause a problem
when we are looking for a specific time (e.g. t=1000), but the actual recorded
timestamp is slighthly off (1000.00001). Round the timestamp based on the
assumption that dt is specified with a limited number of decimal places, and
the intended multiples must have a limited number of decimal places too.