NextEventQueue: better handle reactions which would take population below 0
updatePopulation() already tested for that, but it works on the level
of individual species. Run the test *before* starting updating populations.
This way if a reaction shouldn't happen, or should happen to a lesser
extent, all species will be updated using the same extent.

This should be extended to diffusion and the check in updatePopulation()
turned into an assert.