NextEventQueue: fix bug where reactions with large copy counts were messed up
This bug didn't manifest itself as often as one would expect. The
bug was only triggered when negative (backward) leaps were executed.
Normally, the reaction with higher propensity is the "forward" reaction,
and leap sizes are almost always positive or close to zero. Even if
the propensity ratios flipped, this would usually happen slowly, and
after a sequence of "forward" leaps, there would be some exact events,
and only then a sequence of "backward" leaps. Executing at least one
exact event allowed the reactions to "flip", i.e. be exchanged in the priority
queue. In effect, the leaps were again positive, with the other reaction
being the "forward" one. So for this bug to manifest itself, it was necessary
to have leaps which flipped from fairly high "forward" to fairly high "backward"
leaps. In practice, the number of molecules had to be fairly high (10^6 or so).