NextEventQueue: consider backward and forward reaction leap limits separately
The difference of forward and backward propensities is used when looking
at changes to dependent propensities. The changes to propensity of forward
and backward reactions, iff the latter is present, are considered separately.

The first thing is bugfix. The propensity was used to calculate the allowed
leap size, and by using only the forward propensity, we were almost always
picking a smaller leap size than we could. A too large leap could have
been taken if the reverse propensity was higher then the forward one, but
as established before, this happens only rarely. The actual leap extent was
calculated using both propensities, so this change should have an effect
mostly on speed.

The second change might not be necessary. Need to think this through.