NextEventQueue: do not set reverse propensity to 0
Reverse reaction was not taken into account properly, because the
propensity was artificially set to zero. It was set to zero to move the
reverse event to the bottom of the priority queue. But we can do that
without changing the propensity, by setting the time to infinity. This
way forward and reverse jumps should happen with the right propensity.