Produce package-info.class files
Rename all pkg-info files to package-info and specify package.
It seems that javac puts them in the wrong directory when the package
is not specified, and also javac -Xpkginfo:always does not work with
pkg-info, but only with package-info.

Also rename the ABOUT directory, paths are not case insensitive,
so the package name did not match the path.

Tell the compiler to generate always generate the class files.