NextEventQueue: optimize dependency graph calculation
We'd iterate over the list of all events for every event. While
this works fine for small models, for large models it became quite noticable.
Relace this quadratic behaviour with semi-linear one where the time is
proportional to the number of events times the average number of events in
a voxel.

For Model_SPNspineCaMKII_DagL_AChm4R_long.xml, the time to create the
grah is reduced from 14 min to 9 s.