NextEventQueue: update diffusion variance formula for asymmetrical diffusion
The formula is an approximation based on the first two terms
of the sqrt power series for p1.

It seems that this causes problems when we are near equilibrium:
14:33:33.136 W n.n.g.NextEventQueue - Diffusion CaOut el.213→214, extent 35697 (µ=-102634.24744897336, pop=[1607]):
        max change fraction 1.9038400000000002 @ 99.13784912994521
        for Diffusion CaOut el.214→219 (pop=[54447])
        reverse Diffusion CaOut el.219→214 (pop=[18957])

The problem is most likely that we use the real logarithmic formula when
calculating the leap limit, and then the linear approximation when
calculating the leap size. For diffusion events with large number of
particles and near the equilibrium the first gives us a very large
maximum leap, while the second grossly overestimates the actual number
of firings.