NextEventQueue: downgrade warning level for leaps which are randomly large
We really only care about cases where we calculated the change
fraction in a way that it is too large when the leap is about the
expected size. When we generate a Poisson number that is large,
we'll also get a larger fractianal change than we'd like, but this
should be very infrequent and not a reason to worry.