Rewrite event statistics collection
Instead of separate arrays for diffusion, reaction, and stimulation events,
a uniform array is used. It contains the same information for each event
type: {events, firings}.

The advantage is that we don't have to handle three separate arrays.
It should also be easier to change the amount of information that is
stored in the future.

Only AdaptiveGridCalc implements the new interface, since now the collection
is connected to events, which are only well defined there.