NextEventQueue: fix splitting of injection between voxels
Previous way seems overcomplicated. Let's just distinguish two cases:
submembrane injections which should be split proportionally to the surface area,
and non-submembrane injections which should be split proportionally to the
volume. If we are injecting into a single element, both versions are equivalent,
so there's no need to distinguish this case.