python/docs: update for latest format changes
In ec58e01dd65 dependency writing was disabled, restore it by
the switch in .xml.

Now both the "source" and "target" voxel numbers are included
for the sake of diffusion, so update the format string to only print
the source.

"Next" prefix was removed for the sake of brevity.

make doctest now passes again.