Fix hang at end of simulation when no statistics were written
An empty array of statistics was returned when there were no statistics
to report. The function to create HDF5 tried to increase the chunk
dimensions to be large enough, falling into an infinite loop.

Fix this by not returning an empty array, and also add a few tests here
and there.