StepGenerator: fix generation of leaps with p>1
When doing large leaps for 0 order reactions, we'd attemp an infinite leap
(in practice until the end of the simulation), because there are no constraints
on the leap. For degradation reactions we could try to consume more molecules
than available, resulting in p>1. We should just consume all the molecules in
this case.