Fix SurfaceDensitySet without region
The README said that SurfaceDensitySet can be used without a region,
but this could have never worked. It simply wasn't implemented.

The implementation mirrors what is done for ConcentrationSets, except
that if nothing is specified, we use NaNs. This is becuase for surface
densities, if not specified, we fall back to the value specified by
ConcentrationSets.