NextEventQueue: allow stepper class to be overridden
When making injections with propensity that was slightly below 1 (when multiplied by
the injection duration), the mean number of injected molecules was too low.
This manifested itself when the number of voxels being injected to was
changed, while keeping the summary injection rate constant:
with more voxels, the total number of injected molecules went down.
This looked like the injection rate is subdivided wrong, but actually it
wasn't, and the issue is in the binomial generator. When the generator
is switched to poisson, there is no discrepancy.