Import Binomial and Poisson number generators from colt project
They are not used for anything yet, except that StepGenerator
is extended to allow testing.

Tests show that this code is about as fast as previous code
(InterpolatingStepGenerator), while giving real binomial numbers
and poisson numbers. Distributions match expected values nicely.

numpy.random.binomial is still quite a bit faster. I'm not sure
if this is because of a different algorithm, or because the native
code runs faster. Comparing raw random number generation between
our MersenneTwister and the MersenneTwister used by numpy shows that
numpy is a bit more than twice as fast. This suggests that this
really is a difference between native and non-native code, possibly
in operations like log / exp.