xml: properly detect unexpected character content
The code was written to detect and silence certain kind of errors
(namespace-less elements), so it would store a parsing exception and
only throw it later on, when called again with another parsing exception.
This worked in some cases, but not for the case where one element
had unexpected children. This is now detected properly:

13:54:29.502 E n.x.ModelReader - /mnt/meshugaas/home/zbyszek/neuro/stochastic-test-set/injection/newinjection-borked.xml:10:42: : cvc-complex-type.2.3: Element 'StimulationSet' cannot have character [children], because the type's content type is element-only.