NextEventQueue: treat any injection event as leap barrier
Previous "fix" wasn't enough. In the test simulation, injection of Ca
and leaping of CaBuf combined to create unwanted jumps. With this wider
condition, the issue is avoided.

Unfortunately in the test case, simulation time increases, 4 s → 7 s.
But this is a special example, that only has three molecules and lots
of stimulation events. If the slow down for "real" models proves to be
significant, we can try to make the condition a bit narrower again, e.g.
by limiting it to events which are dependent on another event dependent
on the stimulation (dependency children and grandchildren).